various things
- bug fixes
- supprDoc : check if emprunte
- UI : CSS tweaks
- everything working

+++ src/persistantdata/MediatekData.java
@@ -11,19 +11,20 @@ import java.util.List;
 
 import mediatek2021.*;
 
-// classe mono-instance : l'unique instance est connue de la bibliotheque
-// via une injection de dependance dans son bloc static
-
+/**
+ * 
+ * @see PersistentMediatek
+ */
 public class MediatekData implements PersistentMediatek {
 	private static String url = "jdbc:mysql://localhost:3306/jee?serverTimezone=UTC";
 	private static String user = "root";
 	private static String mdp = "root";
 
+	private static Object lock = new Object();
+
 	private Connection connect;
 
-// Jean-Franeois Brette 01/01/2018
 	static {
-		// injection dynamique de la dependance dans le package stable mediatek2021
 		Mediatek.getInstance().setData(new MediatekData());
 	}
 
@@ -36,34 +37,40 @@ public class MediatekData implements PersistentMediatek {
 		}
 	}
 
-	// renvoie la liste de tous les documents de la bibliotheque
+	/**
+	 * Retourne la liste de tout les documents d'un type donné
+	 * 
+	 * @param type : type de document
+	 * @return la liste des documents du type
+	 */
 	@Override
 	public List<Document> catalogue(int type) {
 		List<Document> documents = new ArrayList<>();
 		String getDocQuery = "SELECT * FROM document D, $table T WHERE D.id_document = T.id_document";
-
-		// TODO : CPO-AV friendly
 		try {
 			switch (type) {
 			case 1: {
 				getDocQuery = getDocQuery.replace("$table", "livre");
 				ResultSet res = connect.createStatement().executeQuery(getDocQuery);
 				while (res.next())
-					documents.add(new Livre(res.getInt("id"), res.getString("titre"), res.getString("auteur"), res.getBoolean("emprunte")));
+					documents.add(new Livre(res.getInt("id"), res.getString("titre"), res.getString("auteur"),
+							res.getBoolean("emprunte")));
 				break;
 			}
 			case 2: {
 				getDocQuery = getDocQuery.replace("$table", "cd");
 				ResultSet res = connect.createStatement().executeQuery(getDocQuery);
 				while (res.next())
-					documents.add(new CD(res.getInt("id"), res.getString("titre"), res.getString("artiste"), res.getBoolean("emprunte")));
+					documents.add(new CD(res.getInt("id"), res.getString("titre"), res.getString("artiste"),
+							res.getBoolean("emprunte")));
 				break;
 			}
 			case 3: {
 				getDocQuery = getDocQuery.replace("$table", "dvd");
 				ResultSet res = connect.createStatement().executeQuery(getDocQuery);
 				while (res.next())
-					documents.add(new DVD(res.getInt("id"), res.getString("titre"), res.getString("realisateur"), res.getBoolean("emprunte")));
+					documents.add(new DVD(res.getInt("id"), res.getString("titre"), res.getString("realisateur"),
+							res.getBoolean("emprunte")));
 				break;
 			}
 			default:
@@ -77,12 +84,19 @@ public class MediatekData implements PersistentMediatek {
 		return documents;
 	}
 
-	// va recuperer le User dans la BD et le renvoie
-	// si pas trouve, renvoie null
+	/**
+	 * Retourne l'utilisateur à partir d'un login et d'un mot de passe
+	 * 
+	 * @param login : identifiant
+	 * @param password : mot de passe
+	 * 
+	 * @return l'utisateur si trouvé, null sinon
+	 */
 	@Override
 	public Utilisateur getUser(String login, String password) {
 		try {
-			PreparedStatement sql = connect.prepareStatement("SELECT * FROM utilisateur WHERE login = ? AND password = ?");
+			PreparedStatement sql = connect
+					.prepareStatement("SELECT * FROM utilisateur WHERE login = ? AND password = ?");
 			sql.setString(1, login);
 			sql.setString(2, password);
 			ResultSet res = sql.executeQuery();
@@ -95,50 +109,53 @@ public class MediatekData implements PersistentMediatek {
 		return null;
 	}
 
-	// va recuperer le document de numero numDocument dans la BD
-	// et le renvoie
-	// si pas trouve, renvoie null
+	/**
+	 * Retourne un document correspond au numéro donné en paramètre
+	 * 
+	 * @param numDocument : le numéro du doc recherché
+	 * 
+	 * @return le document si trouvé, null sinon
+	 */
 	@Override
 	public Document getDocument(int numDocument) {
-		// TODO : CPO-AV friendly
 		String getDocQuery = "SELECT * FROM document D, $table T WHERE D.id_document = $idDoc AND D.id_document = T.id_document";
 		getDocQuery = getDocQuery.replace("$idDoc", Integer.toString(numDocument));
 		try {
 			ResultSet res1 = connect.createStatement().executeQuery(getDocQuery.replace("$table", "livre"));
 			if (res1.next())
-				return new Livre(res1.getInt("id"), res1.getString("titre"), res1.getString("auteur"), res1.getBoolean("emprunte"));
+				return new Livre(res1.getInt("id"), res1.getString("titre"), res1.getString("auteur"),
+						res1.getBoolean("emprunte"));
 
 			ResultSet res2 = connect.createStatement().executeQuery(getDocQuery.replace("$table", "cd"));
 			if (res2.next())
-				return new CD(res2.getInt("id"), res2.getString("titre"), res2.getString("artiste"), res2.getBoolean("emprunte"));
+				return new CD(res2.getInt("id"), res2.getString("titre"), res2.getString("artiste"),
+						res2.getBoolean("emprunte"));
 
 			ResultSet res3 = connect.createStatement().executeQuery(getDocQuery.replace("$table", "dvd"));
 			if (res3.next())
-				return new DVD(res3.getInt("id"), res3.getString("titre"), res3.getString("realisateur"), res3.getBoolean("emprunte"));
+				return new DVD(res3.getInt("id"), res3.getString("titre"), res3.getString("realisateur"),
+						res3.getBoolean("emprunte"));
 		} catch (SQLException e) {
 			System.err.println(e.getMessage());
 		}
 		return null;
 	}
 
-	// ajoute un nouveau document - exception a definir
+	/**
+	 * Ajout un document à la BD selon son type
+	 * 
+	 * @param type : le type de document (1-livre, 2-cd, 3-dvd)
+	 * @param args : un ensemble de propriété propre à chaque type
+	 */
 	@Override
 	public void newDocument(int type, Object... args) throws NewDocException {
-		// args[0] -> le titre
-		// args [1] --> l'auteur
-		// etc en fonction du type et des infos optionnelles
-
-		// TODO : CPO-AV friendly
-		// TODO : check if doc existe déjà
 		try (PreparedStatement docStat = connect.prepareStatement("INSERT INTO document (titre) VALUES (?)",
 				Statement.RETURN_GENERATED_KEYS)) {
-			if (args[0] == null)
-				throw new NewDocException("Aucun titre");
 			docStat.setString(1, (String) args[0]);
 			docStat.execute();
 			ResultSet res = docStat.getGeneratedKeys();
 			if (!res.next())
-				throw new SQLException("Pas de clé générée");
+				throw new SQLException("Erreur : pas de clé générée");
 
 			PreparedStatement typeStatement = null;
 			switch (type) {
@@ -148,7 +165,9 @@ public class MediatekData implements PersistentMediatek {
 				typeStatement.setLong(1, res.getLong(1));
 				typeStatement.setLong(2, res.getLong(1));
 				if (args[1] == null)
-					throw new NewDocException("Pas d'auteur");
+					throw new NewDocException("Merci d'entrer un auteur");
+				if (args[1].equals(""))
+					throw new NewDocException("Merci d'entrer un auteur");
 				typeStatement.setString(3, (String) args[1]);
 				break;
 			case 2:
@@ -156,7 +175,9 @@ public class MediatekData implements PersistentMediatek {
 				typeStatement.setLong(1, res.getLong(1));
 				typeStatement.setLong(2, res.getLong(1));
 				if (args[1] == null)
-					throw new NewDocException("Pas d'artiste");
+					throw new NewDocException("Merci d'entrer un artiste");
+				if (args[1].equals(""))
+					throw new NewDocException("Merci d'entrer un artiste");
 				typeStatement.setString(3, (String) args[1]);
 				break;
 			case 3:
@@ -165,7 +186,9 @@ public class MediatekData implements PersistentMediatek {
 				typeStatement.setLong(1, res.getLong(1));
 				typeStatement.setLong(2, res.getLong(1));
 				if (args[1] == null)
-					throw new NewDocException("Pas de realisateur");
+					throw new NewDocException("Merci d'entrer un realisateur");
+				if (args[1].equals(""))
+					throw new NewDocException("Merci d'entrer un realisateur");
 				typeStatement.setString(3, (String) args[1]);
 				break;
 			default:
@@ -179,15 +202,28 @@ public class MediatekData implements PersistentMediatek {
 
 	}
 
-	// supprime un document - exception a definir
+	/**
+	 * Supprimer un document de la BD selon son id
+	 * 
+	 * @param numDoc le numéro du document à supprimer
+	 */
 	@Override
 	public void suppressDoc(int numDoc) throws SuppressException {
-		//TODO : gerer if emprunte
 		try {
-			PreparedStatement sql = connect.prepareStatement("DELETE FROM document WHERE id_document = ?");
-			sql.setInt(1, numDoc);
-			if (sql.executeUpdate() == 0)
-				throw new SuppressException("Ce numero de document n'existe pas");
+			synchronized (lock) {
+				ResultSet res = connect.createStatement()
+						.executeQuery("SELECT emprunte FROM document WHERE id_document =" + numDoc);
+				if (res.next()) {
+					if (res.getBoolean("emprunte"))
+						throw new SuppressException("Ce document ne peut pas être supprimé, il est emprunté");
+					PreparedStatement sql = connect.prepareStatement("DELETE FROM document WHERE id_document = ?");
+					sql.setInt(1, numDoc);
+					if (sql.executeUpdate() == 0)
+						throw new SuppressException("Il y a eu une erreur lors de la suppression");
+				} else {
+					throw new SuppressException("Ce numero de document n'existe pas");
+				}
+			}
 		} catch (SQLException e) {
 			throw new SuppressException(e.getMessage());
 		}
